Move `retrieve_pull_request_assignments` to `pr_tracking`
Because it needs to use custom logic for determining if a PR is truly assigned to someone and waiting for a review.
